requirements.txt
beautifulsoup4==4.14.2
chromadb==1.3.3
dj_rest_auth==7.0.1
Django==5.2.8
django_allauth==65.12.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
langchain_core==1.0.3
langgraph==1.0.2
matplotlib==3.10.7
networkx==3.4.2
pyautogui==0.9.54
pycryptodome==3.23.0
pydantic==2.12.4
python-dotenv==1.2.1
PyYAML==6.0.3
Requests==2.32.5
selenium==4.38.0
sentence_transformers==5.1.1
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.15.0
gunicorn
psycopg2-binary
celery
redis
mysqlclient
PyMySQL
uvicorn[standard]==0.34.0
httpx==0.28.1 
httpx[http2]
aiomysql==0.3.2